feat: move replay button to canvas toolbar
